+++ dnssec.go
@@ -6,14 +6,11 @@ package main
 
 import (
 	"os"
-	"sync"
 	"time"
 
 	"github.com/miekg/dns"
 	"github.com/skynetservices/skydns/cache"
 )
-
-var inflight = new(single)
 
 // ParseKeyFile read a DNSSEC keyfile as generated by dnssec-keygen or other
 // utilities. It add ".key" for the public key and ".private" for the private key.
@@ -160,43 +157,4 @@ func rrSets(rrs []dns.RR) map[rrset][]dns.RR {
 		return m
 	}
 	return nil
-}
-
-// Adapted from singleinflight.go from the original Go Code. Copyright 2013 The Go Authors.
-type call struct {
-	wg   sync.WaitGroup
-	val  *dns.RRSIG
-	err  error
-	dups int
-}
-
-type single struct {
-	sync.Mutex
-	m map[string]*call
-}
-
-func (g *single) Do(key string, fn func() (*dns.RRSIG, error)) (*dns.RRSIG, error, bool) {
-	g.Lock()
-	if g.m == nil {
-		g.m = make(map[string]*call)
-	}
-	if c, ok := g.m[key]; ok {
-		c.dups++
-		g.Unlock()
-		c.wg.Wait()
-		return c.val, c.err, true
-	}
-	c := new(call)
-	c.wg.Add(1)
-	g.m[key] = c
-	g.Unlock()
-
-	c.val, c.err = fn()
-	c.wg.Done()
-
-	g.Lock()
-	delete(g.m, key)
-	g.Unlock()
-
-	return c.val, c.err, c.dups > 0
 }
